+++ src/Ext/Building/Body.cpp
@@ -10,17 +10,18 @@ void BuildingExt::ExtData::DisplayIncomeString()
 {
 	if (this->AccumulatedIncome && Unsorted::CurrentFrame % 15 == 0)
 	{
-		auto const pOwnerObject = this->OwnerObject();
+		auto const pThis = this->OwnerObject();
 		auto const pTypeExt = this->TypeExtData;
 
-		if ((RulesExt::Global()->DisplayIncome_AllowAI || pOwnerObject->Owner->IsControlledByHuman())
+		if ((RulesExt::Global()->DisplayIncome_AllowAI || pThis->Owner->IsControlledByHuman())
 			&& pTypeExt->DisplayIncome.Get(RulesExt::Global()->DisplayIncome))
 		{
 			FlyingStrings::AddMoneyString(
 				this->AccumulatedIncome,
-				pOwnerObject->Owner,
+				pThis,
+				pThis->Owner,
 				pTypeExt->DisplayIncome_Houses.Get(RulesExt::Global()->DisplayIncome_Houses.Get()),
-				pOwnerObject->GetRenderCoords(),
+				pThis->GetRenderCoords(),
 				pTypeExt->DisplayIncome_Offset
 			);
 		}
@@ -286,7 +287,7 @@ void BuildingExt::ExtData::ApplyPoweredKillSpawns()
 	auto const pThis = this->OwnerObject();
 	auto const pTypeExt = this->TypeExtData;
 
-	if (pTypeExt->Powered_KillSpawns && pThis->Type->Powered && !pThis->IsPowerOnline())
+	if (pTypeExt->Powered_KillSpawns && !pThis->IsPowerOnline())
 	{
 		if (auto const pManager = pThis->SpawnManager)
 		{
@@ -306,7 +307,8 @@ void BuildingExt::ExtData::ApplyPoweredKillSpawns()
 
 bool BuildingExt::ExtData::HandleInfiltrate(HouseClass* pInfiltratorHouse, int moneybefore)
 {
-	const auto pVictimHouse = this->OwnerObject()->Owner;
+	const auto pThis = this->OwnerObject();
+	const auto pVictimHouse = pThis->Owner;
 	const auto pTypeExt = this->TypeExtData;
 	this->AccumulatedIncome += pVictimHouse->Available_Money() - moneybefore;
 
@@ -315,9 +317,10 @@ bool BuildingExt::ExtData::HandleInfiltrate(HouseClass* pInfiltratorHouse, int m
 		// TODO there should be a better way...
 		FlyingStrings::AddMoneyString(
 				this->AccumulatedIncome,
+				pThis,
 				pVictimHouse,
 				pTypeExt->DisplayIncome_Houses.Get(RulesExt::Global()->DisplayIncome_Houses.Get()),
-				this->OwnerObject()->GetRenderCoords(),
+				pThis->GetRenderCoords(),
 				pTypeExt->DisplayIncome_Offset
 		);
 	}
@@ -576,8 +579,7 @@ DEFINE_HOOK(0x454244, BuildingClass_Save_Suffix, 0x7)
 // Removes setting otherwise unused field (0x6FC) in BuildingClass when building has airstrike applied on it so that it can safely be used to store BuildingExt pointer.
 DEFINE_JUMP(LJMP, 0x41D9FB, 0x41DA05);
 
-
-void __fastcall BuildingClass_InfiltratedBy_Wrapper(BuildingClass* pThis, void*, HouseClass* pInfiltratorHouse)
+static void __fastcall BuildingClass_InfiltratedBy_Wrapper(BuildingClass* pThis, void*, HouseClass* pInfiltratorHouse)
 {
 	const int oldBalance = pThis->Owner->Available_Money();
 	// explicitly call because Ares rewrote it