+++ src/Ext/Bullet/Hooks.DetonateLogics.cpp
@@ -94,9 +94,9 @@ DEFINE_HOOK(0x4690C1, BulletClass_Logics_DetonateOnAllMapObjects, 0x8)
 			return vec;
 		};
 
-		auto tryDetonate = [pThis, pWHExt, pOwner, isFull](TechnoClass* pTechno)
+		auto tryDetonate = [pThis, pWHExt, pOwner, isFull](TechnoClass* pTechno, TechnoTypeClass* pType)
 			{
-				if (pWHExt->EligibleForFullMapDetonation(pTechno, pOwner))
+				if (pWHExt->EligibleForFullMapDetonation(pTechno, pType, pOwner))
 				{
 					if (isFull)
 					{
@@ -117,7 +117,7 @@ DEFINE_HOOK(0x4690C1, BulletClass_Logics_DetonateOnAllMapObjects, 0x8)
 			auto const aircraft = copy_dvc(AircraftClass::Array);
 
 			for (auto const pAircraft : aircraft)
-				tryDetonate(pAircraft);
+				tryDetonate(pAircraft, pAircraft->Type);
 		}
 
 		if ((pWHExt->DetonateOnAllMapObjects_AffectTargets & AffectedTarget::Building) != AffectedTarget::None)
@@ -125,7 +125,7 @@ DEFINE_HOOK(0x4690C1, BulletClass_Logics_DetonateOnAllMapObjects, 0x8)
 			auto const buildings = copy_dvc(BuildingClass::Array);
 
 			for (auto const pBuilding : buildings)
-				tryDetonate(pBuilding);
+				tryDetonate(pBuilding, pBuilding->Type);
 		}
 
 		if ((pWHExt->DetonateOnAllMapObjects_AffectTargets & AffectedTarget::Infantry) != AffectedTarget::None)
@@ -133,7 +133,7 @@ DEFINE_HOOK(0x4690C1, BulletClass_Logics_DetonateOnAllMapObjects, 0x8)
 			auto const infantry = copy_dvc(InfantryClass::Array);
 
 			for (auto const pInf : infantry)
-				tryDetonate(pInf);
+				tryDetonate(pInf, pInf->Type);
 		}
 
 		if ((pWHExt->DetonateOnAllMapObjects_AffectTargets & AffectedTarget::Unit) != AffectedTarget::None)
@@ -141,7 +141,7 @@ DEFINE_HOOK(0x4690C1, BulletClass_Logics_DetonateOnAllMapObjects, 0x8)
 			auto const units = copy_dvc(UnitClass::Array);
 
 			for (auto const pUnit : units)
-				tryDetonate(pUnit);
+				tryDetonate(pUnit, pUnit->Type);
 		}
 
 		pThis->Target = pOriginalTarget;
@@ -429,7 +429,7 @@ DEFINE_HOOK(0x469AA4, BulletClass_Logics_Extras, 0x5)
 			if (pTypeExt->ReturnWeapon_ApplyFirepowerMult)
 				damage = static_cast<int>(damage * TechnoExt::GetCurrentFirepowerMultiplier(pTechno));
 
-			if (BulletClass* pBullet = pWeapon->Projectile->CreateBullet(pTechno, pTechno,
+			if (auto const pBullet = pWeapon->Projectile->CreateBullet(pTechno, pTechno,
 				damage, pWeapon->Warhead, pWeapon->Speed, pWeapon->Bright))
 			{
 				BulletExt::SimulatedFiringUnlimbo(pBullet, pOwner, pWeapon, pThis->Location, false);
@@ -819,10 +819,10 @@ DEFINE_HOOK(0x4899DA, MapClass_DamageArea_DamageUnderGround, 0x7)
 
 	auto const pWHExt = WarheadTypeExt::ExtMap.Find(pWH);
 
-	if (!pWHExt || !pWHExt->AffectsUnderground)
+	if (!pWHExt->AffectsUnderground)
 		return 0;
 
-	// bool cylinder = pWHExt->CellSpread_Cylinder;
+	const bool cylinder = pWHExt->CellSpread_Cylinder;
 	const float spread = pWH->CellSpread * (float)Unsorted::LeptonsPerCell;
 
 	for (auto const& pTechno : ScenarioExt::Global()->UndergroundTracker)
@@ -835,9 +835,9 @@ DEFINE_HOOK(0x4899DA, MapClass_DamageArea_DamageUnderGround, 0x7)
 			double dist = 0.0;
 			auto const technoCoords = pTechno->GetCoords();
 
-			//if (cylinder)
-			//	dist = CoordStruct{ technoCoords.X - pCrd->X, technoCoords.Y - pCrd->Y, 0 }.Magnitude();
-			//else
+			if (cylinder)
+				dist = CoordStruct{ technoCoords.X - pCrd->X, technoCoords.Y - pCrd->Y, 0 }.Magnitude();
+			else
 				dist = technoCoords.DistanceFrom(*pCrd);
 
 			if (dist <= spread)