+++ src/Misc/Hooks.Ares.cpp
@@ -16,7 +16,7 @@
 // Patches presented here are exceptions rather that the rule. They must be short, concise and correct.
 // DO NOT POLLUTE ISSUEs and PRs.
 
-ObjectClass* __fastcall CreateInitialPayload(TechnoTypeClass* type, void*, HouseClass* owner)
+static ObjectClass* __fastcall CreateInitialPayload(TechnoTypeClass* type, void*, HouseClass* owner)
 {
 	// temporarily reset the mutex since it's not part of the design
 	const int mutex_old = std::exchange(Unsorted::ScenarioInit, 0);
@@ -25,12 +25,12 @@ ObjectClass* __fastcall CreateInitialPayload(TechnoTypeClass* type, void*, House
 	return instance;
 }
 
-void __fastcall LetGo(TemporalClass* pTemporal)
+static void __fastcall LetGo(TemporalClass* pTemporal)
 {
 	pTemporal->LetGo();
 }
 
-bool __stdcall ConvertToType(TechnoClass* pThis, TechnoTypeClass* pToType)
+static bool __stdcall ConvertToType(TechnoClass* pThis, TechnoTypeClass* pToType)
 {
 	if (const auto pFoot = abstract_cast<FootClass*, true>(pThis))
 		return TechnoExt::ConvertToType(pFoot, pToType);
@@ -39,26 +39,31 @@ bool __stdcall ConvertToType(TechnoClass* pThis, TechnoTypeClass* pToType)
 }
 
 // Technically this replaces GetTechnoType() call.
-TechnoTypeClass* __fastcall ShowPromoteAnim(TechnoClass* pThis)
+static TechnoTypeClass* __fastcall ShowPromoteAnim(TechnoClass* pThis)
 {
 	TechnoExt::ShowPromoteAnim(pThis);
 
 	return pThis->GetTechnoType();
 }
 
-WeaponStruct* __fastcall GetLaserWeapon(BuildingClass* pThis)
+static WeaponStruct* __fastcall GetLaserWeapon(BuildingClass* pThis)
 {
 	return BuildingExt::GetLaserWeapon(pThis);
 }
 
-EBolt* __stdcall CreateEBolt(WeaponTypeClass** pWeaponData)
+static EBolt* __stdcall CreateEBolt(WeaponTypeClass** pWeaponData)
 {
 	return EBoltExt::CreateEBolt(*pWeaponData);
 }
 
-EBolt* __stdcall CreateEBolt2(WeaponTypeClass* pWeapon)
+static EBolt* __stdcall CreateEBolt2(WeaponTypeClass* pWeapon)
 {
 	return EBoltExt::CreateEBolt(pWeapon);
+}
+
+static bool __fastcall CameoIsVeteran(TechnoTypeClass** pTypeExt_Ares, void*, HouseClass* pHouse)
+{
+	return TechnoTypeExt::ExtMap.Find(*pTypeExt_Ares)->CameoIsVeteran(pHouse);
 }
 
 _GET_FUNCTION_ADDRESS(RadarJammerClass::Update, AresRadarJammerClass_Update_GetAddr)
@@ -116,8 +121,11 @@ void Apply_Ares3_0_Patches()
 	// Redirect Ares's RadarJammerClass::Update to our implementation
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x68500, AresRadarJammerClass_Update_GetAddr());
   
-  // Redirect Ares's function to our implementation:
+	// Redirect Ares's function to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x112D0, &BuildingExt::KickOutClone);
+
+	// Redirect Ares's TechnoTypeExt::ExtData::CameoIsElite() to our implementation:
+	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x3D800, &CameoIsVeteran);
 }
 
 void Apply_Ares3_0p1_Patches()
@@ -175,6 +183,9 @@ void Apply_Ares3_0p1_Patches()
 	// Redirect Ares's RadarJammerClass::Update to our implementation
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x69470, AresRadarJammerClass_Update_GetAddr());
   
-  // Redirect Ares's function to our implementation:
+	// Redirect Ares's function to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x11860, &BuildingExt::KickOutClone);
+
+	// Redirect Ares's TechnoTypeExt::ExtData::CameoIsElite() to our implementation:
+	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x3E210, &CameoIsVeteran);
 }